EES-2060 Filter out legend datasets when saving datablocks where the dataset is now orphaned.
